+++ changelog.md
@@ -1,5 +1,9 @@
+0.1.6:
+- arena:
+    - renamed `GrowingArena` -> `Arena`.
+
 0.1.5:
-- slice reader
+- slice reader:
     - `as_slice`, cause deref doesn't return `&'a [T]`.
     - `consume_while_slice[_from]` now return `(&'a [T], bool)`.
     - `consumed_slice`, opposite of `as_slice`.